+++ workspaces/github-actions/packages/app-next/src/App.tsx
@@ -15,15 +15,14 @@
  */
 import { FlatRoutes } from '@backstage/core-app-api';
 import { convertLegacyApp } from '@backstage/core-compat-api';
-import { createApp } from '@backstage/frontend-app-api';
+import { createApp } from '@backstage/frontend-defaults';
 import {
   configApiRef,
-  createApiExtension,
+  ApiBlueprint,
   createApiFactory,
-  createExtensionOverrides,
-  createPageExtension,
-  createSignInPageExtension,
-  createThemeExtension,
+  createFrontendModule,
+  PageBlueprint,
+  SignInPageBlueprint,
 } from '@backstage/frontend-plugin-api';
 import {
   ScmAuth,
@@ -34,17 +33,6 @@ import { ApiExplorerPage } from '@backstage/plugin-api-docs';
 import catalogPlugin from '@backstage/plugin-catalog/alpha';
 import catalogImportPlugin from '@backstage/plugin-catalog-import/alpha';
 import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
-import {
-  createUnifiedTheme,
-  createBaseThemeOptions,
-  pageTheme as defaultPageThemes,
-  palettes,
-  genPageTheme,
-  colorVariants,
-  shapes,
-  UnifiedTheme,
-  UnifiedThemeProvider,
-} from '@backstage/theme';
 import React from 'react';
 import { Navigate, Route } from 'react-router';
 
@@ -53,54 +41,18 @@ import { SignInPage } from './components/auth/SignInPage';
 
 import githubActionsPlugin from '@backstage-community/plugin-github-actions/alpha';
 
-const pageTheme = {
-  ...defaultPageThemes,
-  dataset: genPageTheme({
-    colors: colorVariants.purpleSky,
-    shape: shapes.wave,
-  }),
-};
-
-export const lightTheme: UnifiedTheme = createUnifiedTheme({
-  ...createBaseThemeOptions({
-    palette: { ...palettes.light },
-  }),
-  pageTheme,
+const signInPage = SignInPageBlueprint.make({
+  params: {
+    loader: async () => props => <SignInPage {...props} />,
+  },
 });
 
-export const darkTheme: UnifiedTheme = createUnifiedTheme({
-  ...createBaseThemeOptions({
-    palette: { ...palettes.dark },
-  }),
-  pageTheme,
-});
-
-const lightThemeExtension = createThemeExtension({
-  id: 'mui-light',
-  title: 'MUI Light',
-  variant: 'light',
-  Provider: ({ children }) => (
-    <UnifiedThemeProvider theme={lightTheme} children={children} />
-  ),
-});
-const darkThemeExtension = createThemeExtension({
-  id: 'mui-dark',
-  title: 'MUI Dark',
-  variant: 'dark',
-  Provider: ({ children }) => (
-    <UnifiedThemeProvider theme={darkTheme} children={children} />
-  ),
-});
-
-const signInPage = createSignInPageExtension({
-  name: 'signInPage',
-  loader: async () => props => <SignInPage {...props} />,
-});
-
-const homePageExtension = createPageExtension({
-  namespace: 'home',
-  defaultPath: '/',
-  loader: () => Promise.resolve(<Navigate to="catalog" />),
+const homePageExtension = PageBlueprint.make({
+  name: 'home',
+  params: {
+    defaultPath: '/',
+    loader: () => Promise.resolve(<Navigate to="catalog" />),
+  },
 });
 
 const collectedLegacyPlugins = convertLegacyApp(
@@ -109,16 +61,22 @@ const collectedLegacyPlugins = convertLegacyApp(
   </FlatRoutes>,
 );
 
-const scmAuthExtension = createApiExtension({
-  factory: ScmAuth.createDefaultApiFactory(),
+const scmAuthApi = ApiBlueprint.make({
+  name: 'scm-auth',
+  params: {
+    factory: ScmAuth.createDefaultApiFactory(),
+  },
 });
 
-const scmIntegrationApi = createApiExtension({
-  factory: createApiFactory({
-    api: scmIntegrationsApiRef,
-    deps: { configApi: configApiRef },
-    factory: ({ configApi }) => ScmIntegrationsApi.fromConfig(configApi),
-  }),
+const scmIntegrationsApi = ApiBlueprint.make({
+  name: 'scm-integrations',
+  params: {
+    factory: createApiFactory({
+      api: scmIntegrationsApiRef,
+      deps: { configApi: configApiRef },
+      factory: ({ configApi }) => ScmIntegrationsApi.fromConfig(configApi),
+    }),
+  },
 });
 
 export const app = createApp({
@@ -128,14 +86,13 @@ export const app = createApp({
     userSettingsPlugin,
     githubActionsPlugin,
     ...collectedLegacyPlugins,
-    createExtensionOverrides({
+    createFrontendModule({
+      pluginId: 'app',
       extensions: [
         signInPage,
-        darkThemeExtension,
-        lightThemeExtension,
         homePageExtension,
-        scmAuthExtension,
-        scmIntegrationApi,
+        scmAuthApi,
+        scmIntegrationsApi,
         navigationExtension,
       ],
     }),