+++ workspaces/jenkins/plugins/jenkins/src/alpha/entityCards.test.tsx
@@ -14,14 +14,16 @@
  * limitations under the License.
  */
 import { screen, waitFor } from '@testing-library/react';
-import { createExtensionTester } from '@backstage/frontend-test-utils';
-import * as cards from './entityCards';
 import {
-  createApiExtension,
-  createApiFactory,
-} from '@backstage/frontend-plugin-api';
+  createExtensionTester,
+  renderInTestApp,
+  TestApiProvider,
+} from '@backstage/frontend-test-utils';
+import { EntityProvider } from '@backstage/plugin-catalog-react';
+import * as cards from './entityCards';
 import { JenkinsApi, jenkinsApiRef } from '../api';
 import { sampleEntity } from '../__fixtures__/entity';
+import React from 'react';
 
 jest.mock('@backstage/plugin-catalog-react', () => ({
   ...jest.requireActual('@backstage/plugin-catalog-react'),
@@ -34,33 +36,46 @@ jest.mock('@backstage/core-plugin-api', () => ({
 }));
 
 describe('Entity content extensions', () => {
-  const mockJenkinsApi = createApiExtension({
-    factory: createApiFactory({
-      api: jenkinsApiRef,
-      deps: {},
-      factory: () =>
-        ({
-          getProjects: jest.fn(),
-          getBuild: jest.fn(),
-          getJobBuilds: jest.fn().mockReturnValue({
-            name: 'main',
-            displayName: 'main',
-            description: 'description',
-            fullDisplayName: 'main',
-            inQueue: false,
-            fullName: 'main',
-            url: 'url.com',
-            builds: [],
-          }),
-          retry: () => null,
-        } as unknown as JenkinsApi),
+  const mockJenkinsApi = {
+    getProjects: jest.fn(),
+    getBuild: jest.fn(),
+    getJobBuilds: jest.fn().mockReturnValue({
+      name: 'main',
+      displayName: 'main',
+      description: 'description',
+      fullDisplayName: 'main',
+      inQueue: false,
+      fullName: 'main',
+      url: 'url.com',
+      builds: [],
     }),
-  });
+    retry: () => null,
+  } as unknown as JenkinsApi;
+
+  const mockedEntity = {
+    apiVersion: 'backstage.io/v1alpha1',
+    kind: 'Component',
+    metadata: {
+      name: 'backstage',
+      description: 'backstage.io',
+    },
+    spec: {
+      lifecycle: 'production',
+      type: 'service',
+      owner: 'user:guest',
+    },
+  };
 
   it('should render Jenkins latest run card', async () => {
-    createExtensionTester(cards.entityLatestJenkinsRunCard)
-      .add(mockJenkinsApi)
-      .render();
+    renderInTestApp(
+      <TestApiProvider apis={[[jenkinsApiRef, mockJenkinsApi]]}>
+        <EntityProvider entity={mockedEntity}>
+          {createExtensionTester(
+            cards.entityLatestJenkinsRunCard,
+          ).reactElement()}
+        </EntityProvider>
+      </TestApiProvider>,
+    );
 
     await waitFor(() => {
       expect(screen.getByText('Latest master build')).toBeInTheDocument();
@@ -68,9 +83,13 @@ describe('Entity content extensions', () => {
   });
 
   it('should render Jenkins runs table', async () => {
-    createExtensionTester(cards.entityJobRunsTable)
-      .add(mockJenkinsApi)
-      .render();
+    renderInTestApp(
+      <TestApiProvider apis={[[jenkinsApiRef, mockJenkinsApi]]}>
+        <EntityProvider entity={mockedEntity}>
+          {createExtensionTester(cards.entityJobRunsTable).reactElement()}
+        </EntityProvider>
+      </TestApiProvider>,
+    );
 
     await waitFor(() => {
       expect(screen.getByText('main Runs')).toBeInTheDocument();