+++ workspaces/azure-devops/plugins/azure-devops-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-azure-devops-common
 
+## 0.4.9
+
+### Patch Changes
+
+- dab2f81: Backstage version bump to v1.32.2
+
 ## 0.4.8
 
 ### Patch Changes