+++ workspaces/cicd-statistics/packages/app-next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # app-next
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [a624400]
+  - @backstage-community/plugin-cicd-statistics-module-gitlab@0.1.35
+  - @backstage-community/plugin-cicd-statistics@0.1.41
+
 ## 0.0.5
 
 ### Patch Changes