+++ workspaces/cicd-statistics/plugins/cicd-statistics-module-gitlab/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage-community/plugin-cicd-statistics-module-gitlab
 
+## 0.1.35
+
+### Patch Changes
+
+- a624400: Backstage version bump to v1.32.0
+- Updated dependencies [a624400]
+  - @backstage-community/plugin-cicd-statistics@0.1.41
+
 ## 0.1.34
 
 ### Patch Changes