+++ workspaces/jenkins/packages/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # backend
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [f090030]
+  - @backstage-community/plugin-jenkins-backend@0.6.2
+
 ## 0.0.10
 
 ### Patch Changes