+++ workspaces/jenkins/plugins/jenkins/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage-community/plugin-jenkins
 
+## 0.11.1
+
+### Patch Changes
+
+- f090030: Backstage version bump to v1.32.0
+- Updated dependencies [f090030]
+  - @backstage-community/plugin-jenkins-common@0.1.30
+
 ## 0.11.0
 
 ### Minor Changes