+++ workspaces/redhat-argocd/packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [3865528]
+  - @backstage-community/plugin-redhat-argocd@1.8.7
+
 ## 0.0.5
 
 ### Patch Changes