+++ workspaces/sentry/packages/app-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app-next
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [2dedede]
+  - @backstage-community/plugin-sentry@1.0.1
+
 ## 0.0.2
 
 ### Patch Changes