+++ workspaces/sentry/plugins/sentry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-sentry
 
+## 1.0.1
+
+### Patch Changes
+
+- 2dedede: Backstage version bump to v1.32.0
+
 ## 1.0.0
 
 ### Major Changes